Used size_t as the type of size, count and index variables.
We should use size_t as the type of size, count and index variables
in general, as it provides additional information to the reader.

Note that for many STL containers use size_t as the size_type.

Review: https://reviews.apache.org/r/50374/

+++ src/slave/containerizer/mesos/isolators/gpu/allocator.cpp
@@ -85,7 +85,7 @@ static Try<set<Gpu>> enumerateGpus(
   if (flags.nvidia_gpu_devices.isSome()) {
     indices = flags.nvidia_gpu_devices.get();
   } else {
-    for (unsigned int i = 0; i < resources.gpus().getOrElse(0); ++i) {
+    for (size_t i = 0; i < resources.gpus().getOrElse(0); ++i) {
       indices.push_back(i);
     }
   }
